[Refactor] Move memory lock into runtime memory module
